# src/main/java/org/barbershop/customer/adapter/in/rest/CustomerRequestDTO.java
package org.barbershop.customer.adapter.in.rest;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class CustomerRequestDTO {
    private String name;
    private String phone;
    private String email;
    private String address;
}