# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Barber Shop API
  version: 1.0.0
  description: API para el catálogo, gestión de barbería y ventas.
servers:
  - url: /api
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [Customers]
      summary: Listar clientes con paginación
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
          description: Buscar por nombre o teléfono
        - name: page
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
          description: Número de página (1-indexed)
        - name: pageSize
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 100 }
          description: Cantidad de items por página (máximo 100)
      responses:
        '200':
          description: Lista de clientes con paginación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedCustomers' }
    post:
      operationId: createCustomer
      tags: [Customers]
      summary: Crear un nuevo cliente
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/CustomerRequest' } }
      responses:
        '201':
          description: Cliente creado
          content:
            application/json: { schema: { $ref: '#/components/schemas/Customer' } }
        '400':
          $ref: '#/components/responses/BadRequest'
  /customers/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: getCustomer
      tags: [Customers]
      summary: Obtener un cliente por ID
      responses:
        '200': { description: Cliente, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateCustomer
      tags: [Customers]
      summary: Actualizar un cliente
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CustomerRequest' } } }
      responses:
        '200': { description: Cliente actualizado, content: { application/json: { schema: { $ref: '#/components/schemas/Customer' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      operationId: deleteCustomer
      tags: [Customers]
      summary: Eliminar un cliente
      responses:
        '204': { description: Cliente eliminado }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [Appointments]
      summary: Listar citas con paginación y filtros
      parameters:
        - name: startDate
          in: query
          required: false
          schema: { type: string, format: date }
          description: Filtrar desde fecha
        - name: endDate
          in: query
          required: false
          schema: { type: string, format: date }
          description: Filtrar hasta fecha
        - name: employeeId
          in: query
          required: false
          schema: { type: integer, format: int64 }
          description: Filtrar por empleado
        - name: customerId
          in: query
          required: false
          schema: { type: integer, format: int64 }
          description: Filtrar por cliente
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED] }
          description: Filtrar por estado
        - name: page
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
          description: Número de página (1-indexed)
        - name: pageSize
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 100 }
          description: Cantidad de items por página (máximo 100)
      responses:
        '200':
          description: Lista de citas con paginación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedAppointments' }
    post:
      operationId: createAppointment
      tags: [Appointments]
      summary: Crear una nueva cita
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/AppointmentRequest' } }
      responses:
        '201':
          description: Cita creada
          content:
            application/json: { schema: { $ref: '#/components/schemas/Appointment' } }
        '400':
          $ref: '#/components/responses/BadRequest'
  /appointments/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      operationId: getAppointment
      tags: [Appointments]
      summary: Obtener una cita por ID
      responses:
        '200': { description: Cita, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateAppointment
      tags: [Appointments]
      summary: Actualizar una cita
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentRequest' } } }
      responses:
        '200': { description: Cita actualizada, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      operationId: deleteAppointment
      tags: [Appointments]
      summary: Eliminar una cita
      responses:
        '204': { description: Cita eliminada }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [AuditLogs]
      summary: Obtener logs de auditoría con paginación
      parameters:
        - name: entityType
          in: query
          required: false
          schema: { type: string }
          description: Filtrar por tipo de entidad
        - name: action
          in: query
          required: false
          schema: { type: string, enum: [CREATE, UPDATE, DELETE] }
          description: Filtrar por acción
        - name: page
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
          description: Número de página (1-indexed)
        - name: pageSize
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 100 }
          description: Cantidad de items por página (máximo 100)
      responses:
        '200':
          description: Lista de logs de auditoría
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedAuditLogs' }
components:
  responses:
    NotFound:
      description: Recurso no encontrado
    BadRequest:
      description: Solicitud inválida
  schemas:
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages, hasNextPage]
      properties:
        page:
          type: integer
          description: Número de página actual
        pageSize:
          type: integer
          description: Items por página
        total:
          type: integer
          description: Total de items
        totalPages:
          type: integer
          description: Total de páginas
        hasNextPage:
          type: boolean
          description: Si existe siguiente página
    
    # Items
...
    PaginatedCustomers:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
        pagination:
          $ref: '#/components/schemas/Pagination'
    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerRequest'
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: integer, format: int64, readOnly: true }
            createdAt: { type: string, format: date-time, readOnly: true }
    CustomerRequest:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, maxLength: 100, minLength: 1 }
        phone: { type: string, maxLength: 20 }
        email: { type: string, maxLength: 100, format: email }
        address: { type: string, maxLength: 255 }
    
    # Employees
...
    PaginatedAppointments:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Appointment' }
        pagination:
          $ref: '#/components/schemas/Pagination'
    Appointment:
      allOf:
        - $ref: '#/components/schemas/AppointmentRequest'
        - type: object
          required: [id, createdAt]
          properties:
            id: { type: integer, format: int64, readOnly: true }
            createdAt: { type: string, format: date-time, readOnly: true }
    AppointmentRequest:
      type: object
      required: [customerId, employeeId, scheduledAt]
      properties:
        customerId: { type: integer, format: int64 }
        employeeId: { type: integer, format: int64 }
        scheduledAt: { type: string, format: date-time }
        notes: { type: string, maxLength: 500 }
        status: { type: string, enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED], default: SCHEDULED }
    
    # Sales
...
    PaginatedAuditLogs:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AuditLog' }
        pagination:
          $ref: '#/components/schemas/Pagination'
    AuditLog:
      type: object
      required: [id, entityType, entityId, action, timestamp]
      properties:
        id: { type: integer, format: int64, readOnly: true }
        entityType: { type: string }
        entityId: { type: integer, format: int64 }
        action: { type: string, enum: [CREATE, UPDATE, DELETE] }
        oldValues: { type: object, additionalProperties: true }
        newValues: { type: object, additionalProperties: true }
        userName: { type: string }
        timestamp: { type: string, format: date-time }